requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
ipykernel
missingno
Jinja2
notebook
xgboost
black
python-dotenv
python-box
PyYAML
ensure
joblib
#-e.